+++ src/main/webapp/WEB-INF/views/jsp/customer_details.jsp
@@ -80,12 +80,17 @@ legend {
 				<h2>Customers Details</h2>
 			</div>
 			<div class="col-md-4 add_customer">
-				<button type="submit" class="btn btn-primary"
-					onclick="location.href='<%=request.getContextPath()%>/web/customerWeb/editCustomerForm/${customer.customerID}';">
-					Modify Customer</button>
-				<button type="submit" class="btn btn-primary"
-					onclick="location.href='<%=request.getContextPath()%>/web/customerWeb/delete/${customer.customerID}';">
-					Delete Customer</button>
+				<% if(resourcePermIDs.contains(ResourcePermissionEnum.CUSTOMER_UPDATE.getResourcePermissionID())) { %>
+					<button type="submit" class="btn btn-primary"
+						onclick="location.href='<%=request.getContextPath()%>/web/customerWeb/editCustomerForm/${customer.customerID}';">
+						Modify Customer</button>
+				<% } %>
+				
+				<% if(resourcePermIDs.contains(ResourcePermissionEnum.CUSTOMER_DELETE.getResourcePermissionID())) { %>
+					<button type="submit" class="btn btn-primary"
+						onclick="location.href='<%=request.getContextPath()%>/web/customerWeb/delete/${customer.customerID}';">
+						Delete Customer</button>
+				<% } %>
 			</div>
 		</div>
 		<div class="row top-height">
